+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Changed
+- Bootstrap load as a local asset (was loaded from CDN)
+
 ## 0.5 - 2019-11-21
 
 ### Added